created separate app.jsx
